+++ packages/remax-cli/src/__tests__/fixtures/babelrc/expected/pages/index.js
@@ -181,18 +181,47 @@ for (var prop in isUnitlessNumber) {
   _loop_1(prop);
 }
 
-var __spreadArrays = undefined && undefined.__spreadArrays || function () {
-  for (var s = 0, i = 0, il = arguments.length; i < il; i++) {
-    s += arguments[i].length;
+var __read = undefined && undefined.__read || function (o, n) {
+  var m = typeof Symbol === "function" && o[Symbol.iterator];
+
+  if (!m) {
+    return o;
   }
 
-  for (var r = Array(s), k = 0, i = 0; i < il; i++) {
-    for (var a = arguments[i], j = 0, jl = a.length; j < jl; j++, k++) {
-      r[k] = a[j];
+  var i = m.call(o),
+      r,
+      ar = [],
+      e;
+
+  try {
+    while ((n === void 0 || n-- > 0) && !(r = i.next()).done) {
+      ar.push(r.value);
+    }
+  } catch (error) {
+    e = {
+      error: error
+    };
+  } finally {
+    try {
+      if (r && !r.done && (m = i["return"])) {
+        m.call(i);
+      }
+    } finally {
+      if (e) {
+        throw e.error;
+      }
     }
   }
 
-  return r;
+  return ar;
+};
+
+var __spread = undefined && undefined.__spread || function () {
+  for (var ar = [], i = 0; i < arguments.length; i++) {
+    ar = ar.concat(__read(arguments[i]));
+  }
+
+  return ar;
 };
 var vendorPrefixes = ['webkit', 'moz', 'ms', 'o'];
 
@@ -242,7 +271,7 @@ var plainStyle = function plainStyle(style) {
       value = value + 'rpx';
     }
 
-    return __spreadArrays(acc, [transformReactStyleKey(key) + ":" + ( transformPx(value) ) + ";"]);
+    return __spread(acc, [transformReactStyleKey(key) + ":" + ( transformPx(value) ) + ";"]);
   }, []).join('');
 };
 
@@ -1068,18 +1097,47 @@ function stopPropagation(node) {
   validate(node);
 }
 
-var __spreadArrays$1 = undefined && undefined.__spreadArrays || function () {
-  for (var s = 0, i = 0, il = arguments.length; i < il; i++) {
-    s += arguments[i].length;
+var __read$1 = undefined && undefined.__read || function (o, n) {
+  var m = typeof Symbol === "function" && o[Symbol.iterator];
+
+  if (!m) {
+    return o;
   }
 
-  for (var r = Array(s), k = 0, i = 0; i < il; i++) {
-    for (var a = arguments[i], j = 0, jl = a.length; j < jl; j++, k++) {
-      r[k] = a[j];
+  var i = m.call(o),
+      r,
+      ar = [],
+      e;
+
+  try {
+    while ((n === void 0 || n-- > 0) && !(r = i.next()).done) {
+      ar.push(r.value);
+    }
+  } catch (error) {
+    e = {
+      error: error
+    };
+  } finally {
+    try {
+      if (r && !r.done && (m = i["return"])) {
+        m.call(i);
+      }
+    } finally {
+      if (e) {
+        throw e.error;
+      }
     }
   }
 
-  return r;
+  return ar;
+};
+
+var __spread$1 = undefined && undefined.__spread || function () {
+  for (var ar = [], i = 0; i < arguments.length; i++) {
+    ar = ar.concat(__read$1(arguments[i]));
+  }
+
+  return ar;
 };
 
 function isSyntheticType(inputType) {
@@ -1120,7 +1178,7 @@ function createCallbackProxy(eventType, node, callback) {
       return;
     }
 
-    return callback.apply(void 0, __spreadArrays$1([syntheticEvent], restParams));
+    return callback.apply(void 0, __spread$1([syntheticEvent], restParams));
   };
 }
 
@@ -1220,25 +1278,70 @@ function diffProperties(lastRawProps, nextRawProps) {
   return updatePayload;
 }
 
+var __values = undefined && undefined.__values || function (o) {
+  var s = typeof Symbol === "function" && Symbol.iterator,
+      m = s && o[s],
+      i = 0;
+
+  if (m) {
+    return m.call(o);
+  }
+
+  if (o && typeof o.length === "number") {
+    return {
+      next: function next() {
+        if (o && i >= o.length) {
+          o = void 0;
+        }
+
+        return {
+          value: o && o[i++],
+          done: !o
+        };
+      }
+    };
+  }
+
+  throw new TypeError(s ? "Object is not iterable." : "Symbol.iterator is not defined.");
+};
 var scheduleDeferredCallback = scheduler.unstable_scheduleCallback,
     cancelDeferredCallback = scheduler.unstable_cancelCallback,
     shouldYield = scheduler.unstable_shouldYield,
     now = scheduler.unstable_now;
 
 function processProps(newProps, node, id) {
+  var e_1, _a;
+
   var props = {};
 
-  for (var _i = 0, _a = Object.keys(newProps); _i < _a.length; _i++) {
-    var propKey = _a[_i];
+  try {
+    for (var _b = __values(Object.keys(newProps)), _c = _b.next(); !_c.done; _c = _b.next()) {
+      var propKey = _c.value;
 
-    if (typeof newProps[propKey] === 'function') {
-      var contextKey = REMAX_METHOD + "_" + id + "_" + propKey;
-      node.container.createCallback(contextKey, createCallbackProxy(propKey, node, newProps[propKey]));
-      props[propKey] = contextKey;
-    } else if (propKey === 'style') {
-      props[propKey] = newProps[propKey] || '';
-    } else if (propKey === 'children') ; else {
-      props[propKey] = newProps[propKey];
+      if (typeof newProps[propKey] === 'function') {
+        var contextKey = REMAX_METHOD + "_" + id + "_" + propKey;
+        node.container.createCallback(contextKey, createCallbackProxy(propKey, node, newProps[propKey]));
+        props[propKey] = contextKey;
+      } else if (propKey === 'style') {
+        props[propKey] = newProps[propKey] || '';
+      } else if (propKey === 'children') {// pass
+      } else {
+        props[propKey] = newProps[propKey];
+      }
+    }
+  } catch (e_1_1) {
+    e_1 = {
+      error: e_1_1
+    };
+  } finally {
+    try {
+      if (_c && !_c.done && (_a = _b["return"])) {
+        _a.call(_b);
+      }
+    } finally {
+      if (e_1) {
+        throw e_1.error;
+      }
     }
   }
 
@@ -1342,6 +1445,81 @@ var hostConfig = {
 
 var ReactReconcilerInst = ReactReconciler(hostConfig);
 
+var __read$2 = undefined && undefined.__read || function (o, n) {
+  var m = typeof Symbol === "function" && o[Symbol.iterator];
+
+  if (!m) {
+    return o;
+  }
+
+  var i = m.call(o),
+      r,
+      ar = [],
+      e;
+
+  try {
+    while ((n === void 0 || n-- > 0) && !(r = i.next()).done) {
+      ar.push(r.value);
+    }
+  } catch (error) {
+    e = {
+      error: error
+    };
+  } finally {
+    try {
+      if (r && !r.done && (m = i["return"])) {
+        m.call(i);
+      }
+    } finally {
+      if (e) {
+        throw e.error;
+      }
+    }
+  }
+
+  return ar;
+};
+
+var __spread$2 = undefined && undefined.__spread || function () {
+  for (var ar = [], i = 0; i < arguments.length; i++) {
+    ar = ar.concat(__read$2(arguments[i]));
+  }
+
+  return ar;
+};
+/* eslint-disable prefer-rest-params */
+
+
+if (typeof Function.prototype.call === 'undefined') {
+  Function.prototype.call = function (context) {
+    context = context || window;
+    context.fn = this;
+
+    var args = __spread$2(arguments).slice(1);
+
+    var result = context.fn.apply(context, __spread$2(args));
+    delete context.fn;
+    return result;
+  };
+}
+
+if (typeof Function.prototype.apply === 'undefined') {
+  Function.prototype.apply = function (context) {
+    context = context || window;
+    context.fn = this;
+    var result;
+
+    if (arguments[1]) {
+      result = context.fn.apply(context, __spread$2(arguments[1]));
+    } else {
+      result = context.fn();
+    }
+
+    delete context.fn;
+    return result;
+  };
+}
+
 function isClassComponent(Component) {
   return Component.prototype && typeof Component.prototype.render === 'function';
 }
@@ -1437,6 +1615,49 @@ var __extends = undefined && undefined.__extends || function () {
   };
 }();
 
+var __read$3 = undefined && undefined.__read || function (o, n) {
+  var m = typeof Symbol === "function" && o[Symbol.iterator];
+
+  if (!m) {
+    return o;
+  }
+
+  var i = m.call(o),
+      r,
+      ar = [],
+      e;
+
+  try {
+    while ((n === void 0 || n-- > 0) && !(r = i.next()).done) {
+      ar.push(r.value);
+    }
+  } catch (error) {
+    e = {
+      error: error
+    };
+  } finally {
+    try {
+      if (r && !r.done && (m = i["return"])) {
+        m.call(i);
+      }
+    } finally {
+      if (e) {
+        throw e.error;
+      }
+    }
+  }
+
+  return ar;
+};
+
+var __spread$3 = undefined && undefined.__spread || function () {
+  for (var ar = [], i = 0; i < arguments.length; i++) {
+    ar = ar.concat(__read$3(arguments[i]));
+  }
+
+  return ar;
+};
+
 var DefaultAppComponent =
 /** @class */
 function (_super) {
@@ -1483,18 +1704,47 @@ var __extends$1 = undefined && undefined.__extends || function () {
   };
 }();
 
-var __spreadArrays$2 = undefined && undefined.__spreadArrays || function () {
-  for (var s = 0, i = 0, il = arguments.length; i < il; i++) {
-    s += arguments[i].length;
+var __read$4 = undefined && undefined.__read || function (o, n) {
+  var m = typeof Symbol === "function" && o[Symbol.iterator];
+
+  if (!m) {
+    return o;
   }
 
-  for (var r = Array(s), k = 0, i = 0; i < il; i++) {
-    for (var a = arguments[i], j = 0, jl = a.length; j < jl; j++, k++) {
-      r[k] = a[j];
+  var i = m.call(o),
+      r,
+      ar = [],
+      e;
+
+  try {
+    while ((n === void 0 || n-- > 0) && !(r = i.next()).done) {
+      ar.push(r.value);
+    }
+  } catch (error) {
+    e = {
+      error: error
+    };
+  } finally {
+    try {
+      if (r && !r.done && (m = i["return"])) {
+        m.call(i);
+      }
+    } finally {
+      if (e) {
+        throw e.error;
+      }
     }
   }
 
-  return r;
+  return ar;
+};
+
+var __spread$4 = undefined && undefined.__spread || function () {
+  for (var ar = [], i = 0; i < arguments.length; i++) {
+    ar = ar.concat(__read$4(arguments[i]));
+  }
+
+  return ar;
 };
 function createPageWrapper(Page, query) {
   return (
@@ -1521,7 +1771,7 @@ function createPageWrapper(Page, query) {
               args[_i] = arguments[_i];
             }
 
-            return _this.callLifecycle.apply(_this, __spreadArrays$2([phase], args));
+            return _this.callLifecycle.apply(_this, __spread$4([phase], args));
           };
         };
 
@@ -1543,7 +1793,7 @@ function createPageWrapper(Page, query) {
         var callback = callbackName(phase);
 
         if (this.pageComponentInstance && typeof this.pageComponentInstance[callback] === 'function') {
-          return (_a = this.pageComponentInstance)[callback].apply(_a, args);
+          return (_a = this.pageComponentInstance)[callback].apply(_a, __spread$4(args));
         }
       };
 
@@ -1644,18 +1894,47 @@ var Platform = {
 
 };
 
-var __spreadArrays$3 = undefined && undefined.__spreadArrays || function () {
-  for (var s = 0, i = 0, il = arguments.length; i < il; i++) {
-    s += arguments[i].length;
+var __read$5 = undefined && undefined.__read || function (o, n) {
+  var m = typeof Symbol === "function" && o[Symbol.iterator];
+
+  if (!m) {
+    return o;
   }
 
-  for (var r = Array(s), k = 0, i = 0; i < il; i++) {
-    for (var a = arguments[i], j = 0, jl = a.length; j < jl; j++, k++) {
-      r[k] = a[j];
+  var i = m.call(o),
+      r,
+      ar = [],
+      e;
+
+  try {
+    while ((n === void 0 || n-- > 0) && !(r = i.next()).done) {
+      ar.push(r.value);
+    }
+  } catch (error) {
+    e = {
+      error: error
+    };
+  } finally {
+    try {
+      if (r && !r.done && (m = i["return"])) {
+        m.call(i);
+      }
+    } finally {
+      if (e) {
+        throw e.error;
+      }
     }
   }
 
-  return r;
+  return ar;
+};
+
+var __spread$5 = undefined && undefined.__spread || function () {
+  for (var ar = [], i = 0; i < arguments.length; i++) {
+    ar = ar.concat(__read$5(arguments[i]));
+  }
+
+  return ar;
 };
 
 var Container =
@@ -1735,7 +2014,7 @@ function () {
             };
           }
 
-          _this.context.$spliceData((_a = {}, _a[update.path] = __spreadArrays$3([update.start, update.deleteCount], update.items), _a), callback);
+          _this.context.$spliceData((_a = {}, _a[update.path] = __spread$5([update.start, update.deleteCount], update.items), _a), callback);
         };
 
         var _r = [];
@@ -1828,13 +2107,59 @@ function createPortal(children, containerInfo, key) {
   };
 }
 
+var __read$6 = undefined && undefined.__read || function (o, n) {
+  var m = typeof Symbol === "function" && o[Symbol.iterator];
+
+  if (!m) {
+    return o;
+  }
+
+  var i = m.call(o),
+      r,
+      ar = [],
+      e;
+
+  try {
+    while ((n === void 0 || n-- > 0) && !(r = i.next()).done) {
+      ar.push(r.value);
+    }
+  } catch (error) {
+    e = {
+      error: error
+    };
+  } finally {
+    try {
+      if (r && !r.done && (m = i["return"])) {
+        m.call(i);
+      }
+    } finally {
+      if (e) {
+        throw e.error;
+      }
+    }
+  }
+
+  return ar;
+};
+
+var __spread$6 = undefined && undefined.__spread || function () {
+  for (var ar = [], i = 0; i < arguments.length; i++) {
+    ar = ar.concat(__read$6(arguments[i]));
+  }
+
+  return ar;
+};
 var idCounter = 0;
+
+function generatePageId() {
+  var id = idCounter;
+  var pageId = 'page_' + id;
+  idCounter += 1;
+  return pageId;
+} // for testing
 function createPageConfig(Page) {
   var app = getApp();
-  var id = idCounter;
-  idCounter += 1;
   var config = {
-    pageId: 'page_' + id,
     data: {
       action: {},
       root: {
@@ -1845,6 +2170,7 @@ function createPageConfig(Page) {
     lifecycleCallback: {},
     onLoad: function onLoad(query) {
       var PageWrapper = createPageWrapper(Page, query);
+      this.pageId = generatePageId();
       this.query = query;
       this.container = new Container(this);
       this.element = createPortal(React.createElement(PageWrapper, {
@@ -1887,7 +2213,7 @@ function createPageConfig(Page) {
       var _a2 = callbacks;
 
       var _f = function _f(callback) {
-        result = callback.apply(void 0, args);
+        result = callback.apply(void 0, __spread$6(args));
       };
 
       for (var _i2 = 0; _i2 < _a2.length; _i2++) {
@@ -1901,7 +2227,7 @@ function createPageConfig(Page) {
       var callback = callbackName(lifecycle);
 
       if (this.wrapperRef.current && this.wrapperRef.current[callback]) {
-        return (_a = this.wrapperRef.current)[callback].apply(_a, args);
+        return (_a = this.wrapperRef.current)[callback].apply(_a, __spread$6(args));
       }
     },
     onShow: function onShow() {